Split FPPA state machine into a separate file
